add form login

+++ src/main/java/mjr/personalfinance/api/account/AccountNotFoundException.java
@@ -1,10 +1,10 @@
 package mjr.personalfinance.api.account;
 
-class AccountNotFoundException extends RuntimeException { 
+class AccountNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 7020945971273897041L;
 
-	AccountNotFoundException(Long id) {
+	AccountNotFoundException(Integer id) {
 		super("Could not find account: " + id);
 	}
 }
